+++ ChangeLog.md
@@ -1,5 +1,6 @@
 # v0.4.? (?)
 * Interpolation of the CSF is a bit faster now (thanks to Dongyeon)
+* Added `--result-detailed` argument for saving per-channel, per-frame and per-band visual differences. See `docs/result_detailed.md`.
 
 # v0.4.2 (29/09/2024)
 * Files are now sorted after the wildcard expansion
@@ -8,7 +9,7 @@
 * Added: Support for HLG EOTF (e.g. iPhone HDR video) - thanks to Cosmin Stejerean
 * Added: Processing of videos stored as image frames, described using the C-notation `frame_%04d.png`. New arguments: '--fps' and '--frames'
 * Fixed: A better memory model for estimating how many frames can be processed at once on a GPU. Added '--gpu-mem' argument.
-* Added: 'exposure' field in a display model JSON file.
+* Added: `exposure` field in a display model JSON file.
 * Added: `--result` argument to store results in a CSV file.
 * Fixed: Added examples to README.md and improved documentation.
 * Added: ColorVideoVDP logo.